requirements.txt
streamlit
faster-whisper
PyYAML
